backend/requirements.txt
google-generativeai>=0.8.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
reportlab>=4.2.0
pydantic>=2.9.0
python-multipart>=0.0.9
pillow>=10.4.0

